Add files via upload
This is an example of same funding data that I scraped from the Centraide Montreal website (http://www.centraide-mtl.org/en/agencies/directory-of-agencies-supported-by-centraide/).  I used import.io - a pay-for-service program, but you can use any system to scrape your data as long as it exports to an open data format (eg. .csv, json).