Add files via upload
this is an example of data I scraped from the Centraide Montreal webiste: http://www.centraide-mtl.org/en/agencies/directory-of-agencies-supported-by-centraide/.  I used import.io, a pay-for service program, but you can use any method to scrape data as long as it exports to an open data format (eg. .csv or .json).  

The file still has to be cleaned manually as some of the grants had an extra address field and others did not.